wind and SpatialDependence objects can now read binary files instead of plaintext CSV This is using csvparser.c.  I think this code is about as fast as I/O can be.
Addition of auxillary/ab_convert that converts to-and-from ascii and binary representations of our data

I *think* that this might complete the speeding-of-I/O work.  The only things i can think of is:
 - ability to write binary data from the python code, using the methods already in csvparser, especially for processed wind.
However, i'm not sure whether we'll actually want to do this much.  There is always the ab_convert code that converts between binary/ascii

Refs #41